Add better way to get length of epochs snapshot
Every list entry now also stores the number of following entries, which is relatively cheap to do,
as we only ever append new entries at the start.

This allows us to easily get the current length of the list, as we can simply read the current head
and read how many followers it has.